Encryption and Decryption with Big Blocks (256) Working
Encryption and Decryption with Big Blocks (256) Working!

A phrase will be hidden inside of a block of 256 other random encrypted characters.

Later I plan to add hotwords and other things to tune would-be attackers into random areas.

Later I also plan to add a shuffle to rearrange the data within the stack, as I envisioned it originally in my planning.

Alternatively, I may opt to stick true to the "SPAN" codename and encrypt a character with a whole block and chain multiple together.. more on this later!